ocr-ml-viewer/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv

Pillow
pytesseract
